Remover projeto exibindo o nome do projeto removido, e checagem se o projeto existe, caso não exista, exibir mensagem de erro

+++ app/Http/Controllers/ProjetosController.php
@@ -16,10 +16,12 @@ class ProjetosController extends Controller
             ->orderBy('created_at', 'desc')->get();
 
         $mensagemSucesso = session('mensagem.sucesso');
+        $mensagemErro = session('mensagem.erro');
 
         return view('projetos.index')
             ->with('projetos', $projetos)
-            ->with('mensagemSucesso', $mensagemSucesso);
+            ->with('mensagemSucesso', $mensagemSucesso)
+            ->with('mensagemErro', $mensagemErro);
     }
 
     /**
@@ -36,8 +38,8 @@ class ProjetosController extends Controller
     public function store(Request $request)
     {
         // Mass assignment do Eloquent, para salvar os dados no banco, baseado no atributo fillable, definido na Model
-        Project::create($request->all());
-        session()->flash('mensagem.sucesso', 'Projeto adicionado com sucesso');
+        $projeto = Project::create($request->all());
+        session()->flash('mensagem.sucesso', "Projeto \"{$projeto->titulo}\" adicionado com sucesso");
 
         return to_route('projetos.index');
     }
@@ -71,10 +73,20 @@ class ProjetosController extends Controller
      */
     public function destroy(Request $request)
     {
-        Project::destroy($request->id);
-        // Mensagem de retorno sendo salva na sessão, no modelo flash, para ser apagada após a leitura
-        session()->flash('mensagem.sucesso', 'Projeto removido com sucesso');
+        // Busca do projeto do id enviado
+        $projeto = Project::find($request->id);
 
+        if(is_null($projeto)){
+            // Mensagem de erro sendo salva na sessão, no modelo flash, para ser apagada após a leitura
+            session()->flash('mensagem.erro', "Houve um erro na remoção do projeto");
+            return to_route('projetos.index');
+        }
+
+        // Remoção do projeto
+        Project::destroy($request->id);
+
+        // Mensagem de sucesso sendo salva na sessão, no modelo flash, para ser apagada após a leitura
+        session()->flash('mensagem.sucesso', "Projeto \"{$projeto->titulo}\" removido com sucesso");
         return to_route('projetos.index');
     }
 }
